Refactor the ReverseEngineeringExceptionMessagesHelper type and put it on the proper project.
