+++ NFluent.35.Tests/NFluent.35.Tests.csproj
@@ -95,6 +95,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="EnumerableRelatedTests.cs" />
     <Compile Include="ReadmeRelatedTests.cs" />
+    <Compile Include="ReverseEngineeringExceptionMessagesHelperTests.cs" />
     <Compile Include="StringRelatedTests.cs" />
     <Compile Include="T4\ByteRelatedTests.cs">
       <DependentUpon>NumberTestsGenerator.tt</DependentUpon>
