+++ NFluent.40/NFluent.40.csproj
@@ -167,6 +167,9 @@
     <Compile Include="..\NFluent.35\Extensibility\MessageBlock.cs">
       <Link>Extensibility\MessageBlock.cs</Link>
     </Compile>
+    <Compile Include="..\NFluent.35\Extensibility\ReverseEngineeringExceptionMessagesHelper.cs">
+      <Link>Extensibility\ReverseEngineeringExceptionMessagesHelper.cs</Link>
+    </Compile>
     <Compile Include="..\NFluent.35\Extensibility\ValueBlock.cs">
       <Link>Extensibility\ValueBlock.cs</Link>
     </Compile>
@@ -378,7 +381,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" />
+      <UserProperties ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" />
     </VisualStudio>
   </ProjectExtensions>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
